refactor: admin/application --> application
#2135
