+++ src/main/java/ai/elimu/dao/StoryBookLearningEventDao.java
@@ -4,8 +4,8 @@ import java.util.Calendar;
 import java.util.List;
 import org.springframework.dao.DataAccessException;
 
-import ai.elimu.entity.admin.Application;
 import ai.elimu.entity.analytics.StoryBookLearningEvent;
+import ai.elimu.entity.application.Application;
 import ai.elimu.entity.content.StoryBook;
 
 public interface StoryBookLearningEventDao extends GenericDao<StoryBookLearningEvent> {
