+++ src/main/java/ai/elimu/dao/WordLearningEventDao.java
@@ -3,8 +3,8 @@ package ai.elimu.dao;
 import java.util.Calendar;
 import org.springframework.dao.DataAccessException;
 
-import ai.elimu.entity.admin.Application;
 import ai.elimu.entity.analytics.WordLearningEvent;
+import ai.elimu.entity.application.Application;
 import ai.elimu.entity.content.Word;
 
 public interface WordLearningEventDao extends GenericDao<WordLearningEvent> {
