+++ src/test/java/ai/elimu/dao/ApplicationDaoTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
 
-import ai.elimu.entity.admin.Application;
+import ai.elimu.entity.application.Application;
 
 @SpringJUnitConfig(locations = {
     "file:src/main/webapp/WEB-INF/spring/applicationContext.xml",
